pyproject.toml
[project]
name = "anki-rag"
version = "0.1.0"
description = "Anki RAG 시스템 - 영어 학습 특화 RAG"
authors = [
    {name = "[I"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flagembedding>=1.2",
    "qdrant-client>=1.9",
    "openai>=1.30",
    "pydantic>=2.7",
    "click>=8.1",
    "beautifulsoup4>=4.12",
    "torch>=2.3",
    "python-dotenv>=1.0",
    "transformers>=4.44,<4.52",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "mypy>=1.9",
]
web = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "streamlit>=1.28",
    "python-multipart>=0.0.6",
]
anthropic = ["anthropic>=0.28"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
